testing for get_data_size_to_send

+++ src/http2.c
@@ -783,7 +783,7 @@ int send_headers_or_data_stream_verification(hstates_t *st, uint8_t end_stream){
   return 0;
 }
 
-uint8_t get_size_data_to_send(hstates_t *st){
+uint32_t get_size_data_to_send(hstates_t *st){
     uint32_t available_window = get_window_available_size(st->h2s.outgoing_window);
     if( available_window <= st->hd_lists.data_out_size - st->hd_lists.data_out_sent){
         return available_window;
